requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
pydantic==2.11.3
sqlalchemy==2.0.40
aiosqlite==0.21.0
python-dotenv==1.1.0 